requirements.txt
python-telegram-bot[ext]
fastapi
uvicorn[standard]
requests
apscheduler
sqlalchemy
google-generativeai
python-dotenv
pytesseract
pillow
paho-mqtt
